Move `Widgets.Ui` to its own file
